+++ src/sdk/hub/auth.rs
