+++ src/sdk/hub/auth_client.rs
