+++ src/sdk/mod.rs
@@ -27,13 +27,11 @@ use anyhow::Context;
 use hyper_util::rt::TokioIo;
 
 pub mod attr;
-pub mod auth;
-pub mod auth_client;
 #[cfg(feature = "testcontainers")]
 pub mod container;
+
 pub mod hub;
 pub mod model;
-pub mod store;
 
 pub type Result<T> = anyhow::Result<T>;
